Change boundScopes to object instead of array
There is an exception that occurs when boundScopes is an array, but treated like a dictionary. When angular.foreach iterates over boundScopes when it is an array, there may be many shadow "undefined" elements in the array when an object is added to it.

var test = [];
test[200] = 'value';
alert(test.length);
// result is 201

since boundScopes is treated like a dictionary, its type should be an object.